feat(structure): make all components fully reactive
All of the VRTK prefabs and example components now implement the same
reactive concept that is implemented in Zinnia by using the Malimbe
OnChange handlers to ensure any changes to properties via code or the
inspector adhere to the changed values in real time.

All uses of Lists have been replaced with ObservableLists and every
prefab in VRTK has been updated to use this new mechanism.

This is a huge commit that changes pretty much every file as any sub
chunk of the work would leave errors in the commit.