chore(tests): remove TestHelper scene
The TestHelper scene is no longer required as the Zinnia tests run
in play mode which creates a brand new temporary scene to run the
tests in. The instructions for how to run the tests has also been
removed from the README.md as this is really only a thing for Zinnia.