feat(Interactions): remove moment processor polling for kinematic logic
The FollowAction used a MomentProcessor to poll in the LateUpdate to
see whether to turn on or off the kinematic logic when grabbing and
releasing an interactable.

This doesn't need to be polled with a MomentProcessor as the change of
tracking type already executes a code block, so this kinematic logic
has now been moved to the reactive code block instead which makes
the checking a lot more efficient.